hopitalcomplet

+++ src/Controller/SpecialiteController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 
 class SpecialiteController extends AbstractController
@@ -19,7 +20,14 @@ class SpecialiteController extends AbstractController
     #[Route('/specialite/ajout', name: 'app_specialite_add')]
     public function ajout(EntityManagerInterface $entityManager,Request $request,SpecialiteRepository $specialiteRepository): Response
 
-    {  
+    {  /*  $user=$this->getUser();
+        $roles=['ROLE_ADMIN'];
+   
+       // $roles=$user->getRoles();
+       if (!array_intersect($user->getRoles(), $roles)) {
+         throw new AccessDeniedException('Acces refuse');
+       } */
+       $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $specialite=new Specialite();
 
         $form=$this->createForm(SpecialiteType::class,$specialite);
@@ -47,7 +55,8 @@ class SpecialiteController extends AbstractController
     #[Route('/delete/specialite', name: 'app_specialite_delete')]
     public function delete(EntityManagerInterface $entityManager,Request $request): Response
               
-        {   
+        {  
+           $this->denyAccessUnlessGranted('ROLE_ADMIN');
             $examen=$entityManager->getRepository(Specialite::class)->find(
            ['id'=>$request->request->get('id')]) ;
           $entityManager->remove($examen);
@@ -64,7 +73,8 @@ class SpecialiteController extends AbstractController
         #[Route('/edit/specialite/{id}', name: 'app_specialite_edit')]
         public function edit(EntityManagerInterface $entityManager,Request $request,SpecialiteRepository $specialiteRepository,$id): Response
     
-        { 
+        {  
+           $this->denyAccessUnlessGranted('ROLE_ADMIN');
             $specialite=$specialiteRepository->find($id);
     
             $form=$this->createForm(SpecialiteMiseajouType::class,$specialite);
@@ -86,7 +96,8 @@ class SpecialiteController extends AbstractController
 
         #[Route('/specialite', name: 'app_specialite_liste')]
         public function index(EntityManagerInterface $entity ,Request $request,SpecialiteRepository $specialiteRepository): Response
-        { 
+        {  
+           $this->denyAccessUnlessGranted('ROLE_ADMIN');
             $specialites=$specialiteRepository->findAll();
             return $this->render('specialite/index.html.twig', [
                 'specialites' => $specialites,
